feat: Make restaurant info page responsive
Also improve the footer styles by making them more responsive. The
reason for doing this was because when the padding was on the footer
element, it would cause the footer to be slightly larger than its width
(if the width was set in percentages).

This is also only a slight cosmetic upgrade for the restaurant info
page.